+++ changelog.md
@@ -3,6 +3,7 @@
 - Optimised `HexGridLayer.populate`
 - Updated typescript version
 - Set `additionalProperties` in `getMelvorSchema.cjs` to false
+- Fixed forced aspect ratio of 1 in File Generator Image Preview
 
 # v0.12.0